Fixed compilation for MSVC
- Corrected Windows-specific headers
- Corrected some functions's return types
- Explicitly cast variables to smaller types to silent warnings
- Changed size types in zpack_stream to size_t instead of zpack_u32
- Changed %lu to %llu while formatting zpack_u64 values
- Use strerror_s instead of strerror on Windows

+++ programs/args.c
@@ -31,7 +31,7 @@ zpack_bool args_parse(int argc, char** argv, args_options* options)
         char* arg = argv[i];
         if (arg[0] == '-')
         {
-            int len = strlen(arg);
+            size_t len = strlen(arg);
 
             if (len == 2)
             {
